pipeline/cyber-oss/requirements.txt
beautifulsoup4
requests
PyPDF2
pycryptodome
pycld3